[clang] fix some -Wsign-conversion

+++ src/main/mwrm3_editor.cpp
@@ -103,7 +103,7 @@ class enum_names
         {
             if (name.empty())
             {
-                return &name - names;
+                return static_cast<std::size_t>(&name - names);
             }
         }
         return 0;
